Updating to Season 2 E3

+++ Sparky-core/src/shaders/basic.vert
@@ -2,8 +2,7 @@
 
 layout (location = 0) in vec4 position;
 layout (location = 1) in vec2 uv;
-//layout (location = 2) in float tid;
-layout (location = 2) in int tid;
+layout (location = 2) in float tid;
 layout (location = 3) in vec4 color;
 
 uniform mat4 pr_matrix;
@@ -14,8 +13,7 @@ out DATA
 {
 	vec4 position;
 	vec2 uv;
-	//float tid;
-	int tid;
+	float tid;
 	vec4 color;
 } vs_out;
